refactor: DispatcherServlet 위치 변경
- `techcourse` 패키지에서 `interface21.webmvc` 패키지로 클래스 이동
- `DispatcherServlet` 관련 클래스 및 import 경로 수정
- MVC 구조 및 코드 일관성을 위한 리팩토링
- 불필요한 import 정리
